Elisio/requirements.txt
django
django-enumfields
pylint